thang update code
update code

+++ trainning/src/main/java/com/example/demo/repository/PermissionRepository.java
@@ -1,9 +1,13 @@
 package com.example.demo.repository;
 
 import com.example.demo.entity.Permission;
+import com.example.demo.entity.Roles;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+import java.util.Set;
+
 @Repository
-public interface PermissionRepository extends JpaRepository<Permission, Long> {
+public interface PermissionRepository extends JpaRepository<Permission, Integer> {
 }
